requirements.txt
Django>=3.0,<3.1
djangorestframework>= 3.11,<3.12
psycopg2-binary>=2.8.5,<2.9.0
